gym_backend/requirements.txt
asgiref==3.5.2
Django==4.1.3
django-cors-headers==3.13.0
django-recurrence==1.11.1
djangorestframework==3.14.0
numpy==1.23.5
pandas==1.5.1
python-dateutil==2.8.2
pytz==2022.6
six==1.16.0
sqlparse==0.4.3
tzdata==2022.6
